Use common sidebar instance for GuideArticle and GuideHistory
