+++ frontend/src/pages/guide-history-page.tsx
@@ -1,9 +1,8 @@
 import React from "react";
-import { Container, Text, Flex } from "@mantine/core";
+import { Container, Text } from "@mantine/core";
 import { useGetPage } from "../api/hooks/pages";
 import { useParams } from "react-router";
 import { PageArticleHistory } from "../components/page-article-history";
-import { GuideSideBarLeft } from "../components/guide-sidebar-left";
 
 const GuideHistoryPage: React.FC = () => {
   const { slug } = useParams() as { slug: string };
@@ -17,19 +16,7 @@ const GuideHistoryPage: React.FC = () => {
     );
   }
 
-  return (
-    <Container size="xl" flex={1}>
-      <Flex
-        direction={{ base: "column", sm: "row" }}
-        justify="space-between"
-        align="flex-start"
-        gap="xs"
-      >
-        <GuideSideBarLeft />
-        {page ? <PageArticleHistory key={page.slug} page={page} /> : <div />}
-      </Flex>
-    </Container>
-  );
+  return page ? <PageArticleHistory key={page.slug} page={page} /> : <div />;
 };
 
 export default GuideHistoryPage;
